drain status: move formatter to definition
Formatter was not moved out of drain status manager
